Remove navTrans message in the adapter

+++ algorithms/msgPayloadDef/FilterMsgF32Payload.h
@@ -11,8 +11,8 @@ typedef struct {
     double timeTag;                                       //!< [s]  Current time of validity for output
     int numberOfStates;                                   //!< [-]  Number of scalar states in the state vector
     double covar[MAX_STATES_VECTOR * MAX_STATES_VECTOR];  //!< [-]  Current covariance of the filter
-    float state[MAX_STATES_VECTOR];                       //!< [-]  Current estimated state of the filter
-    float stateError[MAX_STATES_VECTOR];  //!< [-]  Current deviation of the state from the reference state
+    double state[MAX_STATES_VECTOR];                      //!< [-]  Current estimated state of the filter
+    double stateError[MAX_STATES_VECTOR];  //!< [-]  Current deviation of the state from the reference state
 } FilterMsgF32Payload;
 
 #endif
